+++ packages/tasit-action/src/erc721/Estate.js
@@ -1,5 +1,5 @@
 import Contract from "../contract/Contract";
-import estateRegistryABI from "../abi/EstateRegistry.json";
+import estateRegistryABI from "../../../tasit-contracts/abi/EstateRegistry.json";
 
 export default class DecentralandEstate extends Contract {
   constructor(address, wallet) {
