+++ packages/tasit-action/src/erc721/Land.js
@@ -1,5 +1,5 @@
 import Contract from "../contract/Contract";
-import landRegistryABI from "../abi/LANDRegistry.json";
+import landRegistryABI from "../../../tasit-contracts/abi/LANDRegistry.json";
 
 export default class Land extends Contract {
   constructor(address, wallet) {
