+++ packages/tasit-action/src/erc721/NFT.js
@@ -1,5 +1,5 @@
 import Contract from "../contract/Contract";
-import erc721FullABI from "../abi/ERC721Full.json";
+import erc721FullABI from "../../../tasit-contracts/abi/ERC721Full.json";
 
 export default class NFT extends Contract {
   constructor(address, wallet) {
