+++ packages/tasit-contracts/abi/MANAToken.json
